refactor: make all struct to topology
